refactor: 카드 내용 컴포넌트 분리 및 예외 상황 추가
